Some Improvements to better handle wordnet:
Limited support for term_expansion/2:
  - can only call built-in predicates.
  - under top-level term_expansions clauses are passed by consult to pl2wam
    this implies a change to write_pl_state_file (and read_pl_state_file)
    with a new file format (see below).
  - with gplc, the term_expansion/2 clauses must be in the source file (or included)
    but can also be passed via pl-state command-line option.

New pl state file format: new a Prolog file which is loaded (limited to
clause definitions, :- initialization directives and a specific
add_input_term/1 directive to insert a term in the input before the file to
compile is read.

Improve dynamic predicate support (sometimes slow due to the implementation
of Logical Database Update View (LDUV). Some non-deterministic clause
scanning (clause/2, retract/1) run in O(n^2/2) due to successive retractions
(for LDUV, the number of ignored clauses increases one by one). Limit the max
number of increases before physically reclaiming space (clause gc).

Modify pl2wam (read_file.pl) to handle term_expansion and to improve the
compiler for lorge discontiguous predicates.

+++ src/Tools/map_example.c
@@ -142,7 +142,7 @@ example2()
 
   for(int i = 0; t[i]; i++)
     {
-      int len = strlen(t[i]) / 2;
+      int len = (int) strlen(t[i]) / 2;
       sl.str = t[i];
       sl.len = len;
 #ifndef MAP_KEY_PASSED_BY_PTR
